Implemented unique name validation
The forms to create and edit artists and venues now verifies that
the name inputted by the user is unique before validating the form
and displays an error message to the user if the name is not unique.

The logic excepts name matches on existing records so that the edit
form still works as expected by passing the current record's ID to
the forms with a hidden field.

Three additional bug fixes:
    1) The "seeking_description" field now fails validation if the
       user only enters white space.
    2) Removed render_template commands from artist and venue edit
       routes on failure because the template would return as
       undefined
    3) The delete button on the edit_venue.html template no longer
       triggers with the enter key. Implemented the delete button
       on the edit_artist.html template as well.

TODO: add subtitle with record ID to the edit_venue and edit_artist
      templates to match the show_artist and show_venue look and feel.