Added subtitle styling to templates
Added a subtitle style paragraph to the edit_artist.html and
edit_venue.html templates to match the look and feel of the
show_artist.html and show_venue.html templates.

Removed the redundant home button icon from the edit_venue template,
because it does not display correctly if a venu name is long.