Implemented shows/create form and Shows model
Implemented "Shows" model in models.py, and migrated and updated
the database. Implemented logic in the shows/create route in app.py
to update create a show record in the database from the
new_show.html template. Implemented placeholder fields for error
messages to the user in new_show.html. Implemented basic error
messages in ShowForm in forms.py.

TODO: Implement validation in the ShowForm in forms.py to verify
      that the venue ID and artist ID inputed by the user exist
      in the database, and verify that a valid date and time for
      the show is entered by the user.