Implemented validation on ShowForm
The ShowForm in forms.py now validates user input. It verifies that
the artist ID and venue ID provided by the user exists in the
database and returns an error to the user if not. The form also
verifies that a valid date is entered.

CONSIDER: Implement show date restrictions. Presently a user can
          create a show that occurred in the middle ages, or create
          a show that won't happen until long after we're dead.