Implemented shows on artist, venue & venues views
Structured parent-child relationships between Artist and Show, and
Venue and Show database models.

Added query and logic to the /venues view to add number of
upcoming shows to the data passed to the venues.html template.

Added a function to create list of past and upcoming shows, as
well as their counts for the venues/<id> and artists/<id> views
structued so that it can be conditionally added to the data returned
by the 'getRecordAsDict' function and passed to the show_venue.html
and show_artists.html templates.