Implemented shows view, and edit and delete shows
Updated/created routes in app.py to handle viewing the show list,
and editing and deleting show records.

Added a subtitle with the show ID, and edit and delete buttons to
the show tiles on the shows.html template. Updated the height of
the tile-show class in main.css to account for added lines.

Created edit_show.html template to allow existing shows to be
edited.