Refactored venue and artist search function
Refactored the function that returns keyword search results for
venues and artists so that the shows query only runs if the
'shows_match_id' parameter is passed.

Also added conditionals to the artist and venue search controllers
so that the page redirects to the artists page or venues page if
the search field is empty or contains only whitespace.