Implemented recent listings on home pages
The home page now displays the five most recently listed artists
and venues using a similar style to the show tile. The images are
also clickable links.

Created a funnction to return the five most recent (by ID) artists
and venues along with the count (the function should be easy to modify
by changing the limit on the query if more or less listings are
desired). The results are formatted in a dict to be passed by
the controller to the home.html template.

Created a CSS class in main.css to size the tiles appropriately.

CONSIDER: make image links clickable on the show_artist.html and
show_venue.html templates.