Implemented artist availability
User can now enter available times for artists. Changed date inputs
to strings which are formatted to datetime objects when passed to
the database because the WTForms DateTime fields cannot be empty.

Created a custom validator to verify entered date is in a
YYYY-MM-DD HH:MM format and allow empty fields or fields of
white space to pass validation (empty fields and whitespace are
are converted to None types when passed to the database).

Created a validator to verify that an artist's available start date
is before an artist's available end date.

Created a validator that verifies a date entered on the create show
form is in the artist's availability range. That validator also
checks the date entered against an artists existing shows +/-
a configurable hour(s) time delta so that artist's can't have shows
booked too closely together.